requirements.txt
torch
torchvision
torchaudio
pydicom
opencv-python
numpy
pandas
matplotlib
scikit-learn
ultralytics
Pillow
tqdm
seaborn
